move beam search classes
