+++ vllm/engine/async_llm_engine.py
@@ -28,11 +28,11 @@ from vllm.outputs import (CompletionOutput, EmbeddingRequestOutput,
 from vllm.pooling_params import PoolingParams
 from vllm.prompt_adapter.request import PromptAdapterRequest
 from vllm.sampling_params import BeamSearchParams, SamplingParams
-from vllm.sequence import ExecuteModelRequest
+from vllm.sequence import (BeamSearchSequence, ExecuteModelRequest,
+                           create_sort_beams_key_function)
 from vllm.transformers_utils.tokenizer import AnyTokenizer
 from vllm.usage.usage_lib import UsageContext
-from vllm.utils import (BeamSearchSequence, collect_from_async_generator,
-                        create_sort_beams_key_function, deprecate_kwargs,
+from vllm.utils import (collect_from_async_generator, deprecate_kwargs,
                         random_uuid, weak_bind)
 
 logger = init_logger(__name__)
