+++ vllm/entrypoints/llm.py
@@ -23,13 +23,14 @@ from vllm.pooling_params import PoolingParams
 from vllm.prompt_adapter.request import PromptAdapterRequest
 from vllm.sampling_params import (BeamSearchParams, GuidedDecodingParams,
                                   RequestOutputKind, SamplingParams)
+from vllm.sequence import (BeamSearchInstance, BeamSearchOutput,
+                           BeamSearchSequence)
 from vllm.transformers_utils.tokenizer import (AnyTokenizer, MistralTokenizer,
                                                get_cached_tokenizer)
 from vllm.transformers_utils.tokenizer_group import TokenizerGroup
 from vllm.usage.usage_lib import UsageContext
-from vllm.utils import (BeamSearchInstance, BeamSearchOutput,
-                        BeamSearchSequence, Counter, deprecate_kwargs,
-                        get_beam_search_score, is_list_of)
+from vllm.utils import (Counter, deprecate_kwargs, get_beam_search_score,
+                        is_list_of)
 
 logger = init_logger(__name__)
